Important reorganisation of project:
- module basic contains stuff to manage basic stream io::{Read, Write} is reimport at niffler first level
- module seek contains stuff to manage compression support seek
